👋 editorState and rendererState
All data previously stored on the editor and renderer states is stored directly in the store (duh): model, animations, wireframe, vertexSource, uniforms, etc.

Most editor and renderer states data is kept up to date via v-model bindings on the components, but for the uniforms values and vertex/fragment codes the “commitState” is used to update the store values.